Refactor our Chunk interface and start implementing the first chunked v4
