+++ jhdf/src/main/java/io/jhdf/btree/Chunk.java
@@ -1,0 +1,13 @@
+package io.jhdf.btree;
+
+import java.util.BitSet;
+
+public interface Chunk {
+    int getSize();
+
+    BitSet getFilterMask();
+
+    int[] getChunkOffset();
+
+    long getAddress();
+}
